Refactor: Extract screen health hooks into separate modules
This change breaks down the large useScreenHealthIntegration hook into smaller, more manageable hooks for daemon status, events, profiles, and screenshots. This improves code organization and reusability.
